requirements.txt
django
djangorestframework
drf-yasg
django-cors-headers
django-filter
djangorestframework-simplejwt
psycopg2-binary
gspread 
google-auth-oauthlib
dj-rest-auth
django-allauth<0.58.0
python-barcode
Pillow